Fix Internal Bug at cscannerHelp.c:2428.
This bug is supposedly fixed upstream, however in actual fact, the
problem just doesn't manifest itself as long as the ./configure script is
executed with defaults. Only when ./configure is run with --prefix=/usr,
as is normally done in Debian, does it rear its ugly head.

Despite the bug not really being fixed, forcing SYSTEM_LIBDIR to be set
with the default value, at least, gets us the same behaviour as upstream.

Closes: #473595
Closes: #476228
Closes: #498345
Closes: #548768
Closes: #569928
Closes: #601424